Rename versions package to packetManager as it is more suitable
